requirements/development.txt
-r base.txt
pytest==7.4.3
pytest-django==4.7.0
django-debug-toolbar==4.2.0
factory-boy==3.3.0